Add support for Plasma 6
With the release of Plasma 6, there is no longer any direct equivalent
of `clientDemandsAttentionChanged` in the workspace object. Instead, we
have to watch for window creation, and hook the `demandsAttention` event
there for each window. From there, it works basically the same as
before.

The Plasma 5 implementation remains the same as before, the script now starts by
detecting which version it is running on.